Work with JDBC corrected

+++ src/main/java/com/forensic/entity/crime/CrimeBuilder.java
@@ -14,17 +14,17 @@ public interface CrimeBuilder {
     /**
      * Add police registration number for builder
      */
-    CrimeBuilder policeRegNumber(long policeRegNumber);
+    CrimeBuilder policeRegNumber(Long policeRegNumber);
 
     /**
      * Add number of criminal case for builder
      */
-    CrimeBuilder caseInvestigationNumber(long caseInvestigationNumber);
+    CrimeBuilder caseInvestigationNumber(Long caseInvestigationNumber);
 
     /**
      * Add number of Criminal Code article for builder
      */
-    CrimeBuilder criminalCodeArticleNumber(int criminalCodeArticleNumber);
+    CrimeBuilder criminalCodeArticleNumber(Long criminalCodeArticleNumber);
 
     /**
      * Add short description of crime for builder
